fix: #21 Cache 컴포넌트 분류
